fix: rename image files to kebab-case, fix imports
Renames 3 asset files that had spaces (and one typo) in their names:
- src/assets/images/Youth meetup.jpg -> youth-meetup.jpg
- src/assets/images/smart devices.jpg -> smart-devices.jpg (currently unused/unimported)
- src/assets/team/Jean Pual.jpg -> jean-paul.jpg (typo fix: Pual -> Paul)

Updates the two imports that referenced the old filenames
(FeaturedProject.tsx, constants/about.ts). Verified with `npm run build`
that both renamed assets resolve correctly.

Closes #283
