+++ src/assets/images/smart-devices.jpg
